Many more small polishes to use of primitives/objects, etc.
git-svn-id: https://svn.apache.org/repos/asf/lucene/mahout/trunk@704462 13f79535-47bb-0310-9956-ffa450edef68

+++ core/src/main/java/org/apache/mahout/classifier/bayes/common/BayesTfIdfMapper.java
@@ -65,8 +65,8 @@ public class BayesTfIdfMapper extends MapReduceBase implements
         throw new IOException(label);
       }
       
-      Float labelDocumentCount = labelDocumentCounts.get(label);
-      float logIdf = (float)Math.log(labelDocumentCount.floatValue()  / value.get());
+      float labelDocumentCount = labelDocumentCounts.get(label);
+      float logIdf = (float) Math.log(labelDocumentCount  / value.get());
       
       output.collect(new Text(labelFeaturePair), new FloatWritable(logIdf));
     } else if (labelFeaturePair.startsWith(",")) {
